Feature: Removes writing to the disk
The plugin used to write its content to the disk.
This has several issues, for example that you might
overwrite a file on accident.

This commit alters the code such that a scratch
buffer is used which is discarded at the end of
the session. This solves most of the problems of
the file-based approach.